requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
googletrans==4.0.0rc1
yt-dlp==2024.4.9
wikipedia==1.4.0
